* Get rid of many gcc warnings.

+++ main/gbc/gbc_trans.h
@@ -138,7 +138,7 @@ void TRANS_want(int reserved, char *msg);
 void TRANS_want_newline(void);
 //int TRANS_get_class(PATTERN pattern);
 bool TRANS_is_end_function(bool is_proc, PATTERN *look);
-char *TRANS_get_num_desc(int num);
+char *TRANS_get_num_desc(ushort num);
 
 #define TRANS_is(_reserved) (PATTERN_is(*JOB->current, (_reserved)) ? JOB->current++, TRUE : FALSE)
 #define TRANS_ignore(_reserved) (void)TRANS_is(_reserved)
